feat(eKoNLPy): this is a BREAKING CHANGE. (#8)
This is a BREAKING CHANGE. We've made several updates to enhance the functionality, readability, and usability of eKoNLPy. Here are the main points:

Mecab Class Integration: We've merged the extended Mecab class of eKoNLPy with the original Fugashi Mecab class. This aims to streamline the use of both versions and improve code maintenance.

Explicit Type Information: To improve readability and robustness, we've added explicit type information across the codebase. This change facilitates understanding of the data types used and promotes safer coding practices.

Parameter Name Changes: Please note that some parameter names have been modified to reflect their purpose more accurately.

Clarification of Exposed Classes: We've clearly identified the classes exposed from init. This clarification should help users and developers understand the structure and functionality of eKoNLPy.

Command Line Interface (CLI): To enhance usability, we've added a new Command Line Interface. You can now execute a variety of operations directly from the command line, thus making eKoNLPy more user-friendly and accessible.

Please refer to the updated documentation for details on these changes and instructions on how to adjust your code to this new version. We understand this is a breaking change and apologize for any inconvenience. However, we believe these changes will greatly enhance the functionality, readability, and usability of eKoNLPy in the long run.